Added feature to require php files in ns statement.
This commit adds a new feature to load php files in the ns statement. This
can, for example, be used to load the composer autoload file. Example

(ns myapp
  (:require-file "vendor/autload.php"))

The :require-file import ins transformed into a require_once statement in PHP.
The file is loaded before any other :require import.

Fixes: #409