Evaluate meta data in the special form def
This commit evaluates the meta data that can be provided in def. This was
currently not a problem because all data stored in the map was literal data
that evaluates to itself.

Before this commit the code

(def x {:foo (+ 1 1)} 10)

was evaluated to

(def x {:foo '(+ 1 1)} 10) # The + operation is not evaluated

With the commit the code evaluates to

(def x {:foo 2} 10)

This change required some changes to the GlobalEnvironment. The meta data
of a definition is now loaded from the Registry and is not stored in the
global enviornment itself. This has the drawback that the meta data
is only available after the defintion was evaluated and not while analysing the
initial value of the definition. However, this only happens in recursive function
calls. In this case the meta data will be a empty persistent map.