Added support for inline optimization.
This commits adds the possibility to mark functions for inline optimization. Inline
optimization is very useful for math and bit operation. It returns the number of
instruction that need to be executed. For example before this commit

(+ 1 (* 2 3))

would lead two to function calls. First, Phel's *-function is called which itself
calls the PHP's *-function. Then the same happens for the +-function. The inline
optimization would directly call PHP's function. The compiled PHP code looks exactly
like expected:

(1 + (2 * 3))

This can now be archived by adding :inline and optional :inline-arity meta data to a
function. :inline is a macro function that returns new optimized code that should be
analyzed instead of the normal function. Similar to macros.
:inline-arity is an optional precondition that can be added. It is a one argument
function that returns true or false. The value of the argument is amount of arguments
that a analyzed call retrives.