Fixes the compile command when it rebuild using a cache.
The compile command had the problem that the cached version was not really
working. After executing the command the first time everything went fine,
because the was not cached files. In the second run the command mostly fails
because of undefined functions. The problem here is that the GlobalEnvironment
didn't know of functions in the cached files.

This commit fixes the problem by importing the already compiled files.