Docker quickstart container
This commits adds a Dockerfile to build a Docker image that starts the REPL.
A GitHub actions workflow is added to automatically publish a version to docker
hub. New user can use this image to try out Phel very easily. A one line
command will be added to Phel's website.

The command is: docker run -it --rm phellang/repl